requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
pymysql==1.1.1
pydantic-settings==2.8.1
python-dotenv==1.2.1
kimi-cli==1.26.0
pytest==8.3.5
pytest-asyncio==0.25.3
